+++ src/firestarter/OneAPI/OneAPI.cpp
@@ -265,7 +265,7 @@ OneAPI::OneAPI(const volatile firestarter::LoadThreadWorkType& LoadVar, bool Use
 
 void OneAPI::initGpus(GpuFlop& ExecutedFlop, std::condition_variable& WaitForInitCv, std::mutex& WaitForInitCvMutex,
                       bool& InitDone, const volatile firestarter::LoadThreadWorkType& LoadVar, bool UseFloat,
-                      bool UseDouble, unsigned MatrixSize, int Gpus) {
+                      bool UseDouble, uint64_t MatrixSize, int Gpus) {
   std::condition_variable GpuThreadsWaitForInitCv;
   std::mutex GpuThreadsWaitForInitCvMutex;
   std::vector<std::thread> GpuThreads;
